[refactor] 쿼리 위치 분리
