+++ movie/core/src/main/kotlin/com/example/movie/domain/movie/repository/MovieRepository.kt
@@ -1,8 +1,11 @@
 package com.example.movie.domain.movie.repository
 
 import com.example.movie.domain.movie.model.Movie
+import com.example.movie.domain.screening.model.Screening
+import java.time.LocalDateTime
 
 interface MovieRepository {
     fun findById(id: Long): Movie?
     fun findAllOrderByReleaseDateDesc(): List<Movie>
+    fun findAllNowPlayingWithMovieAndTheater(currentTime: LocalDateTime): Map<Movie, List<Screening>>
 }
