+++ movie/core/src/main/kotlin/com/example/movie/service/MovieService.kt
@@ -2,18 +2,18 @@ package com.example.movie.service
 
 import com.example.movie.domain.common.TimeHandler
 import com.example.movie.domain.movie.model.Movie
+import com.example.movie.domain.movie.repository.MovieRepository
 import com.example.movie.domain.screening.model.Screening
-import com.example.movie.domain.screening.repository.ScreeningRepository
 import org.springframework.stereotype.Service
 
 @Service
 class MovieService(
-    private val screeningRepository: ScreeningRepository,
+    private val movieRepository: MovieRepository,
     private val timeHandler: TimeHandler
 ) : NowPlayingMoviesUseCase {
 
     override fun getNowPlayingMovies(): Map<Movie,List<Screening>> {
         val currentTime = timeHandler.getCurrentTime()
-        return screeningRepository.findAllNowPlayingWithMovieAndTheater(currentTime)
+        return movieRepository.findAllNowPlayingWithMovieAndTheater(currentTime)
     }
 }
